refactor(app): create base class to extend when creating PageEffects classes
Since we register effects on a page-by-page basis, it is possible to have strange side effects when
two or more pages are watching the same action to trigger an effect. So we use OnRunEffects callback
by ngrx team to instruct it when our page effects should be run. In this case, they only happen when
the page is the active page.